bedstatus/linux: bat(): use unsigned long

+++ bedstatus/linux.c
@@ -181,7 +181,7 @@ static char *bat_read (const struct bat *bat, const char *name)
 
 static void bat (struct status *st)
 {
-	unsigned nowsum = 0, fullsum = 0;
+	unsigned long nowsum = 0, fullsum = 0;
 	st->has_bat_charging = false;
 
 	for (int i = 0; i < nbat; ++i) {
